first float function

+++ calculator.py
@@ -1,4 +1,4 @@
-x = int(input("What is your first number? "))
-y = int(input("What is your second number? "))
+x = float(input("What is your first number? "))
+y = float(input("What is your second number? "))
 
 print(x + y)
